Add files via upload
Project Description:
The Movie Website project is a dynamic web application created using JavaScript, HTML, and CSS. It utilizes an API to fetch data from a website and displays the top 20 movies worldwide. The website includes a search bar functionality that allows users to find specific movies by typing their names. Additionally, when hovering over a movie poster, a sliding slide appears, providing an overview of the movie.

Technologies Used:

JavaScript: Used to handle the dynamic behavior of the website, including fetching data from the API, implementing the search functionality, and creating the sliding slide effect.
HTML: Provides the structure and layout for the website, including the search bar, movie posters, and sliding slide sections.
CSS: Used for styling and enhancing the visual presentation of the website, including formatting the layout, designing the sliding slide, and applying hover effects.
Features and Functionality:

API Integration: The project utilizes an API to fetch data from a website, specifically obtaining information about the top 20 movies across the globe. This API data is then dynamically displayed on the website.
Search Bar Functionality: Users can search for specific movies by typing their names in the search bar. The website filters the movies based on the search input, displaying the relevant results.
Sliding Slide Effect: When hovering over a movie poster, a sliding slide appears, revealing additional information and an overview of the movie. This effect enhances the user experience and provides quick insights about the movies.
The Movie Website project demonstrates the effective use of JavaScript, HTML, and CSS to create a user-friendly and interactive platform for movie enthusiasts. By integrating an API to fetch data and incorporating features such as the search bar and sliding slide effect, the website offers an engaging and informative movie browsing experience.